fix: add game over script
